refactor league details and login controller. Adds miscellaneous comments, and moves classes into correct place.
