+++ MockStock/Controllers/MSLeagueDetailsViewController.swift
@@ -10,11 +10,14 @@ import Foundation
 import UIKit
 
 class MSLeagueDetailsViewController: UIViewController, UICollectionViewDataSource, UICollectionViewDelegateFlowLayout {
+    
+    // MARK: Data Properties
     var leagueUsers = [LeagueUser]()
     var leagueName = "League Name"
     var leagueId: String?
     var hostId: Int?
     
+    // MARK: View Properties
     var tableView: UITableView = {
         let tv = UITableView()
         tv.translatesAutoresizingMaskIntoConstraints = false
@@ -28,6 +31,7 @@ class MSLeagueDetailsViewController: UIViewController, UICollectionViewDataSourc
         return v
     }()
 
+    // MARK: ViewController Methods
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -58,6 +62,7 @@ class MSLeagueDetailsViewController: UIViewController, UICollectionViewDataSourc
         fetchData()
     }
     
+    // MARK: CollectionView DataSource+Delegate Methods
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
         return leagueUsers.count
     }
@@ -81,6 +86,7 @@ class MSLeagueDetailsViewController: UIViewController, UICollectionViewDataSourc
         return CGSize(width: collectionView.frame.width - 60, height: collectionView.frame.width * 0.15)
     }
     
+    // MARK: API Requests
     func fetchData() {
         guard let lid = leagueId else { return }
         
@@ -198,90 +204,4 @@ class MSLeagueDetailsViewController: UIViewController, UICollectionViewDataSourc
         }.resume() // fires the session
     }
     
-}
-
-class MSLeagueUserCell: UICollectionViewCell {
-    
-    // Background bars
-    var colorView1: UIView = {
-        let v = UIView()
-        v.backgroundColor = UIColor.darkGray
-        return v
-    }()
-    var colorView2: UIView = {
-        let v = UIView()
-        v.backgroundColor = UIColor.darkGray
-        return v
-    }()
-    
-    // Labels
-    var userName: UILabel = {
-        let l = UILabel()
-        l.text = "Someone"
-        l.font = UIFont(name: "Futura-CondensedMedium", size: 24)
-        l.textColor = .white
-        l.textAlignment = .left
-        return l
-    }()
-    var netWorth: UILabel = {
-        let l = UILabel()
-        l.text = "$500"
-        l.font = UIFont(name: "Futura-CondensedMedium", size: 24)
-        l.textColor = .white
-        l.textAlignment = .left
-        return l
-    }()
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setupViews()
-    }
-    
-    func setupViews() {
-        // Configure view properties
-        layer.cornerRadius = self.frame.width / 25
-//        layer.shadowColor = UIColor.black.cgColor
-//        layer.shadowOffset = CGSize(width: 1.0, height: 2.0)
-//        layer.shadowOpacity = 0.4
-//        layer.shadowRadius = 4.0
-        layer.borderColor = UIColor.black.cgColor
-        layer.borderWidth = 4.0
-        clipsToBounds = true
-        
-        // Add & Configure subviews
-        addSubview(colorView1)
-        addSubview(colorView2)
-        
-        addSubview(userName)
-        addSubview(netWorth)
-        
-        netWorth.textAlignment = .right
-        
-    }
-    
-    override func didMoveToSuperview() {
-        super.didMoveToSuperview()
-        
-        // Size background views
-        colorView2.frame = CGRect(x: 0, y: 0, width: frame.width, height: frame.height * 0.5)
-        //colorView2.layer.cornerRadius = colorView2.frame.width / 25
-        colorView1.frame = CGRect(x: 0, y: frame.height * 0.5, width: frame.width, height: frame.height * 0.5)
-        
-        setupLabels()
-    }
-    
-    private func setupLabels() {
-        userName.frame = CGRect(x: 15, y: frame.height/2 - 20, width: frame.width, height: 40)
-        netWorth.frame = CGRect(x: frame.width/2, y: frame.height/2 - 20, width: frame.width/2 - 15, height: 40)
-    }
-    
-    func setColors(topView: UIColor, bottomViewColor: UIColor) {
-        self.backgroundColor = topView
-        self.colorView1.backgroundColor = topView
-        self.colorView2.backgroundColor = bottomViewColor
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
 }
